+++ service/runtime/events.go
@@ -11,9 +11,12 @@ const (
 	// EventServiceUpdated is the topic events are published to when a service is updated
 	EventServiceUpdated = "service.updated"
 	// EventServiceDeleted is the topic events are published to when a service is deleted
-	EventServiceDeleted   = "service.deleted"
-	EventNamespaceCreated = "namespace.created"
-	EventNamespaceDeleted = "namespace.deleted"
+	EventServiceDeleted       = "service.deleted"
+	EventNamespaceCreated     = "namespace.created"
+	EventNamespaceDeleted     = "namespace.deleted"
+	EventNetworkPolicyCreated = "networkpolicy.created"
+	EventNetworkPolicyUpdated = "networkpolicy.updated"
+	EventNetworkPolicyDeleted = "networkpolicy.deleted"
 )
 
 // EventPayload which is published with runtime events
@@ -26,5 +29,12 @@ type EventPayload struct {
 // EventNamespacePayload which is published with runtime namespace events
 type EventNamespacePayload struct {
 	Type      string
+	Namespace string
+}
+
+// EventNetworkPolicyPayload which is published with runtime networkpolicy events
+type EventNetworkPolicyPayload struct {
+	Type      string
+	Name      string
 	Namespace string
 }
